StorageWipeColumn: remove unnecessary properties
Consequently, the subiquity type extensions had to be moved to a
separate file to avoid circular dependencies.
