+++ packages/ubuntu_desktop_installer/lib/pages/allocate_disk_space/allocate_disk_space_model.dart
@@ -1,13 +1,10 @@
 import 'dart:async';
 
-import 'package:filesize/filesize.dart';
 import 'package:flutter/foundation.dart';
 import 'package:subiquity_client/subiquity_client.dart';
 import 'package:ubuntu_wizard/utils.dart';
 
 import '../../services.dart';
-
-export 'package:subiquity_client/subiquity_client.dart' show Disk, Partition;
 
 /// The default mount points for auto-completion.
 const kDefaultMountPoints = <String>[
@@ -262,13 +259,4 @@ class AllocateDiskSpaceModel extends ChangeNotifier {
     _selectionController.close();
     super.dispose();
   }
-}
-
-extension DiskExtension on Disk {
-  String get prettySize => filesize(size ?? 0);
-}
-
-extension PartitionExtension on Partition {
-  bool get canWipe => preserve == true && mount != null;
-  String get prettySize => filesize(size ?? 0);
 }
