+++ packages/ubuntu_desktop_installer/lib/pages/allocate_disk_space/storage_table.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:scroll_to_index/scroll_to_index.dart';
-import 'package:subiquity_client/subiquity_client.dart';
 import 'package:ubuntu_wizard/widgets.dart';
 
 import 'storage_columns.dart';
+import 'storage_types.dart';
 
 typedef CanSelectStorage = bool Function(int disk, [int partition]);
 typedef IsStorageSelected = bool Function(int disk, [int partition]);
